+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".LogIn" />
+        <activity android:name=".Login.LogIn" />
         <activity android:name=".SplashIntegrApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SignUp"
+        <activity android:name=".SignUp.SignUp"
             android:windowSoftInputMode="stateHidden"/>
         <activity
             android:name=".MainActivity"
