+++ app/src/main/java/com/integrapp/integrapp/Chat/ChatAppMsgAdapter.java
@@ -1,7 +1,6 @@
-package com.integrapp.integrapp;
+package com.integrapp.integrapp.Chat;
 
 
-import android.content.Context;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
