+++ app/src/main/java/com/integrapp/integrapp/Forum/OtherFragment.java
@@ -1,4 +1,4 @@
-package com.integrapp.integrapp;
+package com.integrapp.integrapp.Forum;
 
 import android.annotation.SuppressLint;
 import android.os.AsyncTask;
@@ -14,8 +14,8 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.Toast;
 
-import com.integrapp.integrapp.Forum.ForumItem;
-import com.integrapp.integrapp.Forum.ForumsAdapter;
+import com.integrapp.integrapp.R;
+import com.integrapp.integrapp.Server;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -23,7 +23,8 @@ import org.json.JSONObject;
 
 import java.util.ArrayList;
 
-public class LangFragment extends android.support.v4.app.Fragment {
+
+public class OtherFragment extends android.support.v4.app.Fragment {
 
     private Server server;
     private ArrayList<ForumItem> threads = new ArrayList<>();
@@ -37,7 +38,7 @@ public class LangFragment extends android.support.v4.app.Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_forum_lang, container, false);
+        View view = inflater.inflate(R.layout.fragment_forum_other, container, false);
         this.server = Server.getInstance();
         llista = view.findViewById(R.id.llista);
         setInfoForum();
@@ -56,11 +57,11 @@ public class LangFragment extends android.support.v4.app.Fragment {
 
             @Override
             protected String doInBackground(Void... voids) {
-                return server.getForumLang();
+                return server.getForumOther();
             }
 
             protected void onPostExecute(String s) {
-                if (!s.equals("ERROR IN GETTING FORUM(LANGUAGES)")) {
+                if (!s.equals("ERROR IN GETTING FORUM(OTHER)")) {
                     try {
                         getInfoFromString(s);
                         forumsAdapter = new ForumsAdapter(getContext(), threads);
@@ -113,6 +114,7 @@ public class LangFragment extends android.support.v4.app.Fragment {
             if (description.length()> 48) {
                 miniDescription = description.substring(0,48) + "...";
             }
+
             ForumItem item = new ForumItem(id, type, title, description, miniDescription, createdAt, userId, rate, user);
 
             threads.add(item);
