+++ app/src/main/java/com/integrapp/integrapp/Inscription/InscriptionsAdapter.java
@@ -1,12 +1,13 @@
-package com.integrapp.integrapp;
+package com.integrapp.integrapp.Inscription;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
-import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.integrapp.integrapp.R;
 
 import java.util.List;
