+++ app/src/main/java/com/integrapp/integrapp/SplashIntegrApp.java
@@ -3,9 +3,11 @@ package com.integrapp.integrapp;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Bundle;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
+
+import com.integrapp.integrapp.Login.LogIn;
 
 public class SplashIntegrApp extends AppCompatActivity {
