+++ app/src/main/java/com/integrapp/integrapp/User.java
@@ -1,6 +1,6 @@
 package com.integrapp.integrapp;
 
-class User {
+public class User {
     private String id;
     private String username;
     private String name;
