+++ app/src/main/java/com/integrapp/integrapp/UsersFragment.java
@@ -15,9 +15,13 @@ import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import com.integrapp.integrapp.Chat.SingleChatFragment;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+
 import java.util.ArrayList;
 
 public class UsersFragment extends android.support.v4.app.Fragment {
