+++ app/src/main/res/layout/activity_log_in.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="18dp"
     android:paddingRight="18dp"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".LogIn">
+    tools:context=".Login.LogIn">
 
     <LinearLayout
         android:layout_width="match_parent"
