+++ app/src/main/res/layout/activity_new_advert.xml
@@ -11,7 +11,7 @@
         android:paddingLeft="18dp"
         android:paddingRight="18dp"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context="com.integrapp.integrapp.NewAdvertFragment">
+        tools:context="com.integrapp.integrapp.Adverts.NewAdvertFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
