+++ app/src/main/res/layout/activity_profile.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.integrapp.integrapp.ProfileFragment"
+    tools:context="com.integrapp.integrapp.Profile.ProfileFragment"
     android:orientation="vertical">
 
     <RelativeLayout
