+++ app/src/main/res/layout/activity_sign_up.xml
@@ -11,7 +11,7 @@
     android:paddingLeft="18dp"
     android:paddingRight="18dp"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.integrapp.integrapp.SignUp">
+    tools:context="com.integrapp.integrapp.SignUp.SignUp">
 
     <ScrollView
         android:layout_width="match_parent"
