+++ app/src/main/res/layout/single_advert_fragment.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.integrapp.integrapp.SingleAdvertFragment"
+    tools:context="com.integrapp.integrapp.Adverts.SingleAdvertFragment"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="18dp"
     android:paddingRight="18dp"
